restructure the app into pages
