+++ src/App.js
@@ -3,9 +3,9 @@ import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
 import "./App.css";
 import Navbar from "./components/Navbar";
-import About from "./components/About";
-import Portfolio from "./components/Portfolio";
-import Contact from "./components/Contact";
+import About from "./pages/About";
+import Portfolio from "./pages/Portfolio";
+import Contact from "./pages/Contact";
 import CTA from "./components/CTA";
 import  Footer from "./components/Footer";
 
@@ -21,7 +21,6 @@ function App() {
         </Routes>
         <CTA />
        <Footer />
-       
     </Router>
     </div>
   );
